Expand example Salt master configuration
This adds Pillar data for users-formula and for salt-formula's
'salt.formulas' SLS, which together complete the configuration example
of using Salt (and Git) to manage the master configuration itself.

Additionally, this adds some comments explaining what each section of
the Pillar SLS file does.  The comments hint at a bootstrapping
procedure, which ultimately requires manually (and temporarily)
checking out the necessary Git repositories on the master.